+++ crates/wgsl-types/src/builtin/ctor.rs
@@ -1,12 +1,23 @@
-//! Implementations of the constructor functions, including zero-value constructors.
+//! Constructor implementations, including zero-value constructors.
+//!
+//! Functions bear the same name as the WGSL counterpart.
+//! Functions that take template parameters are suffixed with `_t` and take `tplt_*` arguments.
+//!
+//! ### Usage quirks
+//!
+//! * The arguments must be [loaded][Type::loaded].
+//! * The struct constructor is a bit special since it is the only user-defined type.
+//!   Use [`struct_ctor`] and [`typecheck_struct_ctor`] for structs.
+//! * User-defined functions can shadow WGSL built-in functions.
+//! * Type aliases must be resolved: WGSL allows calling functions with the name of the alias.
 
 use half::prelude::*;
 use itertools::Itertools;
-use num_traits::Zero;
+use num_traits::{FromPrimitive, One, ToPrimitive, Zero};
 
 use crate::{
-    CallSignature, Error,
-    conv::{Convert, convert_all_ty},
+    CallSignature, Error, ShaderStage,
+    conv::{Convert, convert_all, convert_all_inner_to, convert_all_to, convert_all_ty},
     inst::{ArrayInstance, Instance, LiteralInstance, MatInstance, StructInstance, VecInstance},
     tplt::{ArrayTemplate, MatTemplate, TpltParam, VecTemplate},
     ty::{StructType, Ty, Type},
@@ -14,6 +25,10 @@ use crate::{
 
 type E = Error;
 
+/// Check if a function name could correspond to a built-in constructor function.
+///
+/// Warning: WGSL allows shadowing built-in functions. Check that a user-defined
+/// function does not shadow the built-in one.
 pub fn is_ctor(name: &str) -> bool {
     matches!(
         name,
@@ -38,10 +53,486 @@ pub fn is_ctor(name: &str) -> bool {
     )
 }
 
-/// Call a struct constructor.
-pub fn struct_ctor(struct_ty: &StructType, args: &[Instance]) -> Result<Instance, E> {
+// ------------
+// CONSTRUCTORS
+// ------------
+// reference: <https://www.w3.org/TR/WGSL/#constructor-builtin-function>
+
+/// `array<T,N>()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#array-builtin>
+pub fn array_t(tplt_ty: &Type, tplt_n: usize, args: &[Instance]) -> Result<Instance, E> {
+    let args = args
+        .iter()
+        .map(|a| {
+            a.convert_to(tplt_ty).ok_or_else(|| {
+                E::ParamType(Type::Array(Box::new(tplt_ty.clone()), Some(tplt_n)), a.ty())
+            })
+        })
+        .collect::<Result<Vec<_>, _>>()?;
+
+    if args.len() != tplt_n {
+        return Err(E::ParamCount("array".to_string(), tplt_n, args.len()));
+    }
+
+    Ok(ArrayInstance::new(args, false).into())
+}
+
+/// `array()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#array-builtin>
+pub fn array(args: &[Instance]) -> Result<Instance, E> {
+    let args = convert_all(args).ok_or(E::Builtin("array elements are incompatible"))?;
+
     if args.is_empty() {
-        return StructInstance::zero_value(struct_ty).map(Instance::from);
+        return Err(E::Builtin("array constructor expects at least 1 argument"));
+    }
+
+    Ok(ArrayInstance::new(args, false).into())
+}
+
+/// `bool()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#bool-builtin>
+pub fn bool(a1: &Instance) -> Result<Instance, E> {
+    match a1 {
+        Instance::Literal(l) => {
+            let zero = LiteralInstance::zero_value(&l.ty())?;
+            Ok(LiteralInstance::Bool(*l != zero).into())
+        }
+        _ => Err(E::Builtin("bool constructor expects a scalar argument")),
+    }
+}
+
+/// `i32()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#i32-builtin>
+// TODO: check that "If T is a floating point type, e is converted to i32, rounding towards zero."
+pub fn i32(a1: &Instance) -> Result<Instance, E> {
+    match a1 {
+        Instance::Literal(l) => {
+            let val = match l {
+                LiteralInstance::Bool(n) => Some(n.then_some(1).unwrap_or(0)),
+                LiteralInstance::AbstractInt(n) => n.to_i32(), // identity if representable
+                LiteralInstance::AbstractFloat(n) => Some(*n as i32), // rounding towards 0
+                LiteralInstance::I32(n) => Some(*n),           // identity operation
+                LiteralInstance::U32(n) => Some(*n as i32),    // reinterpretation of bits
+                LiteralInstance::F32(n) => Some((*n as i32).min(2147483520)), // rounding towards 0 AND representable in f32
+                LiteralInstance::F16(n) => Some((f16::to_f32(*n) as i32).min(65504)), // rounding towards 0 AND representable in f16
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::I64(n) => n.to_i32(), // identity if representable
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::U64(n) => n.to_i32(), // identity if representable
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::F64(n) => Some(*n as i32), // rounding towards 0
+            }
+            .ok_or(E::ConvOverflow(*l, Type::I32))?;
+            Ok(LiteralInstance::I32(val).into())
+        }
+        _ => Err(E::Builtin("i32 constructor expects a scalar argument")),
+    }
+}
+
+/// `u32()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#u32-builtin>
+pub fn u32(a1: &Instance) -> Result<Instance, E> {
+    match a1 {
+        Instance::Literal(l) => {
+            let val = match l {
+                LiteralInstance::Bool(n) => Some(n.then_some(1).unwrap_or(0)),
+                LiteralInstance::AbstractInt(n) => n.to_u32(), // identity if representable
+                LiteralInstance::AbstractFloat(n) => Some(*n as u32), // rounding towards 0
+                LiteralInstance::I32(n) => Some(*n as u32),    // reinterpretation of bits
+                LiteralInstance::U32(n) => Some(*n),           // identity operation
+                LiteralInstance::F32(n) => Some((*n as u32).min(4294967040)), // rounding towards 0 AND representable in f32
+                LiteralInstance::F16(n) => Some((f16::to_f32(*n) as u32).min(65504)), // rounding towards 0 AND representable in f16
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::I64(n) => n.to_u32(), // identity if representable
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::U64(n) => n.to_u32(), // identity if representable
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::F64(n) => Some(*n as u32), // rounding towards 0
+            }
+            .ok_or(E::ConvOverflow(*l, Type::U32))?;
+            Ok(LiteralInstance::U32(val).into())
+        }
+        _ => Err(E::Builtin("u32 constructor expects a scalar argument")),
+    }
+}
+
+/// `f32()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#f32-builtin>
+pub fn f32(a1: &Instance, _stage: ShaderStage) -> Result<Instance, E> {
+    match a1 {
+        Instance::Literal(l) => {
+            let val = match l {
+                LiteralInstance::Bool(n) => Some(n.then_some(f32::one()).unwrap_or(f32::zero())),
+                LiteralInstance::AbstractInt(n) => n.to_f32(), // implicit conversion
+                LiteralInstance::AbstractFloat(n) => n.to_f32(), // implicit conversion
+                LiteralInstance::I32(n) => Some(*n as f32),    // scalar to float (never overflows)
+                LiteralInstance::U32(n) => Some(*n as f32),    // scalar to float (never overflows)
+                LiteralInstance::F32(n) => Some(*n),           // identity operation
+                LiteralInstance::F16(n) => Some(f16::to_f32(*n)), // exactly representable
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::I64(n) => n.to_f32(), // implicit conversion
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::U64(n) => n.to_f32(), // implicit conversion
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::F64(n) => n.to_f32(), // implicit conversion
+            }
+            .ok_or(E::ConvOverflow(*l, Type::F32))?;
+            Ok(LiteralInstance::F32(val).into())
+        }
+        _ => Err(E::Builtin("f32 constructor expects a scalar argument")),
+    }
+}
+
+/// `f16()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#f16-builtin>
+pub fn f16(a1: &Instance, stage: ShaderStage) -> Result<Instance, E> {
+    match a1 {
+        Instance::Literal(l) => {
+            let val = match l {
+                LiteralInstance::Bool(n) => Some(n.then_some(f16::one()).unwrap_or(f16::zero())),
+                LiteralInstance::AbstractInt(n) => {
+                    // scalar to float (can overflow)
+                    if stage == ShaderStage::Const {
+                        let range = -65504..=65504;
+                        range.contains(n).then_some(f16::from_f32(*n as f32))
+                    } else {
+                        Some(f16::from_f32(*n as f32))
+                    }
+                }
+                LiteralInstance::AbstractFloat(n) => {
+                    // scalar to float (can overflow)
+                    if stage == ShaderStage::Const {
+                        let range = -65504.0..=65504.0;
+                        range.contains(n).then_some(f16::from_f32(*n as f32))
+                    } else {
+                        Some(f16::from_f32(*n as f32))
+                    }
+                }
+                LiteralInstance::I32(n) => {
+                    // scalar to float (can overflow)
+                    if stage == ShaderStage::Const {
+                        f16::from_i32(*n)
+                    } else {
+                        Some(f16::from_f32(*n as f32))
+                    }
+                }
+                LiteralInstance::U32(n) => {
+                    // scalar to float (can overflow)
+                    if stage == ShaderStage::Const {
+                        f16::from_u32(*n)
+                    } else {
+                        Some(f16::from_f32(*n as f32))
+                    }
+                }
+                LiteralInstance::F32(n) => {
+                    // scalar to float (can overflow)
+                    if stage == ShaderStage::Const {
+                        let range = -65504.0..=65504.0;
+                        range.contains(n).then_some(f16::from_f32(*n))
+                    } else {
+                        Some(f16::from_f32(*n))
+                    }
+                }
+                LiteralInstance::F16(n) => Some(*n), // identity operation
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::I64(n) => {
+                    // scalar to float (can overflow)
+                    if stage == ShaderStage::Const {
+                        let range = -65504..=65504;
+                        range.contains(n).then_some(f16::from_f32(*n as f32))
+                    } else {
+                        Some(f16::from_f32(*n as f32))
+                    }
+                }
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::U64(n) => {
+                    // scalar to float (can overflow)
+                    if stage == ShaderStage::Const {
+                        f16::from_u64(*n)
+                    } else {
+                        Some(f16::from_f32(*n as f32))
+                    }
+                }
+                #[cfg(feature = "naga_ext")]
+                LiteralInstance::F64(n) => {
+                    // scalar to float (can overflow)
+                    if stage == ShaderStage::Const {
+                        let range = -65504.0..=65504.0;
+                        range.contains(n).then_some(f16::from_f32(*n as f32))
+                    } else {
+                        Some(f16::from_f32(*n as f32))
+                    }
+                }
+            }
+            .ok_or(E::ConvOverflow(*l, Type::F16))?;
+            Ok(LiteralInstance::F16(val).into())
+        }
+        _ => Err(E::Builtin("f16 constructor expects a scalar argument")),
+    }
+}
+
+/// `matCxR<T>()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#mat2x2-builtin>
+pub fn mat_t(
+    c: usize,
+    r: usize,
+    tplt_ty: &Type,
+    args: &[Instance],
+    stage: ShaderStage,
+) -> Result<Instance, E> {
+    // overload 1: mat conversion constructor
+    if let [Instance::Mat(m)] = args {
+        if m.c() != c || m.r() != r {
+            return Err(E::Conversion(
+                m.ty(),
+                Type::Mat(c as u8, r as u8, Box::new(tplt_ty.clone())),
+            ));
+        }
+
+        let conv_fn = match tplt_ty {
+            Type::F32 => f32,
+            Type::F16 => f16,
+            _ => return Err(E::Builtin("matrix type must be a f32 or f16")),
+        };
+
+        let comps = m
+            .iter_cols()
+            .map(|v| {
+                v.unwrap_vec_ref()
+                    .iter()
+                    .map(|n| conv_fn(n, stage))
+                    .collect::<Result<Vec<_>, _>>()
+                    .map(|s| Instance::Vec(VecInstance::new(s)))
+            })
+            .collect::<Result<Vec<_>, _>>()?;
+
+        Ok(MatInstance::from_cols(comps).into())
+    } else {
+        let ty = args
+            .first()
+            .ok_or(E::Builtin("matrix constructor expects arguments"))?
+            .ty();
+        let ty = ty
+            .convert_inner_to(tplt_ty)
+            .ok_or(E::Conversion(ty.inner_ty(), tplt_ty.clone()))?;
+        let args =
+            convert_all_to(args, &ty).ok_or(E::Builtin("matrix components are incompatible"))?;
+
+        // overload 2: mat from column vectors
+        if ty.is_vec() {
+            if args.len() != c as usize {
+                return Err(E::ParamCount(format!("mat{c}x{r}"), c, args.len()));
+            }
+
+            Ok(MatInstance::from_cols(args).into())
+        }
+        // overload 3: mat from float values
+        else if ty.is_float() {
+            if args.len() != c * r {
+                return Err(E::ParamCount(format!("mat{c}x{r}"), c * r, args.len()));
+            }
+
+            let args = args
+                .chunks(r)
+                .map(|v| Instance::Vec(VecInstance::new(v.to_vec())))
+                .collect_vec();
+
+            Ok(MatInstance::from_cols(args).into())
+        } else {
+            Err(E::Builtin(
+                "matrix constructor expects float or vector of float arguments",
+            ))
+        }
+    }
+}
+
+/// `matCxR()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#mat2x2-builtin>
+pub fn mat(c: usize, r: usize, args: &[Instance]) -> Result<Instance, E> {
+    // overload 1: mat conversion constructor
+    if let [Instance::Mat(m)] = args {
+        if m.c() != c || m.r() != r {
+            let ty2 = Type::Mat(c as u8, r as u8, m.inner_ty().into());
+            return Err(E::Conversion(m.ty(), ty2));
+        }
+        // note: `matCxR(e: matCxR<S>) -> matCxR<S>` is no-op
+        Ok(m.clone().into())
+    } else {
+        let tys = args.iter().map(|a| a.ty()).collect_vec();
+        let ty = convert_all_ty(&tys).ok_or(E::Builtin("matrix components are incompatible"))?;
+        let mut inner_ty = ty.inner_ty();
+
+        if inner_ty.is_abstract_int() {
+            // force conversion from AbstractInt to a float type
+            inner_ty = Type::AbstractInt;
+        } else if !inner_ty.is_float() {
+            return Err(E::Builtin(
+                "matrix constructor expects float or vector of float arguments",
+            ));
+        }
+
+        let args = convert_all_inner_to(args, &inner_ty)
+            .ok_or(E::Builtin("matrix components are incompatible"))?;
+
+        // overload 2: mat from column vectors
+        if ty.is_vec() {
+            if args.len() != c {
+                return Err(E::ParamCount(format!("mat{c}x{r}"), c, args.len()));
+            }
+
+            Ok(MatInstance::from_cols(args).into())
+        }
+        // overload 3: mat from float values
+        else if ty.is_float() || ty.is_abstract_int() {
+            if args.len() != c * r {
+                return Err(E::ParamCount(format!("mat{c}x{r}"), c * r, args.len()));
+            }
+            let args = args
+                .chunks(r)
+                .map(|v| Instance::Vec(VecInstance::new(v.to_vec())))
+                .collect_vec();
+
+            Ok(MatInstance::from_cols(args).into())
+        } else {
+            Err(E::Builtin(
+                "matrix constructor expects float or vector of float arguments",
+            ))
+        }
+    }
+}
+
+/// `vecN<T>()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#vec2-builtin>
+pub fn vec_t(
+    n: usize,
+    tplt_ty: &Type,
+    args: &[Instance],
+    stage: ShaderStage,
+) -> Result<Instance, E> {
+    // overload 1: vec init from single scalar value
+    if let [Instance::Literal(l)] = args {
+        let val = l
+            .convert_to(tplt_ty)
+            .map(Instance::Literal)
+            .ok_or_else(|| E::ParamType(tplt_ty.clone(), l.ty()))?;
+        let comps = (0..n).map(|_| val.clone()).collect_vec();
+        Ok(VecInstance::new(comps).into())
+    }
+    // overload 2: vec conversion constructor
+    else if let [Instance::Vec(v)] = args {
+        let ty = Type::Vec(n as u8, Box::new(tplt_ty.clone()));
+        if v.n() != n {
+            return Err(E::Conversion(v.ty(), ty));
+        }
+
+        let conv_fn = match ty.inner_ty() {
+            Type::Bool => |n, _| bool(n),
+            Type::I32 => |n, _| i32(n),
+            Type::U32 => |n, _| u32(n),
+            Type::F32 => |n, stage| f32(n, stage),
+            Type::F16 => |n, stage| f16(n, stage),
+            _ => return Err(E::Builtin("vector type must be a scalar")),
+        };
+
+        let comps = v
+            .iter()
+            .map(|n| conv_fn(n, stage))
+            .collect::<Result<Vec<_>, _>>()?;
+
+        Ok(VecInstance::new(comps).into())
+    }
+    // overload 3: vec init from component values
+    else {
+        // flatten vecN args
+        let args = args
+            .iter()
+            .flat_map(|a| -> Box<dyn Iterator<Item = &Instance>> {
+                match a {
+                    Instance::Vec(v) => Box::new(v.iter()),
+                    _ => Box::new(std::iter::once(a)),
+                }
+            })
+            .collect_vec();
+        if args.len() != n {
+            return Err(E::ParamCount(format!("vec{n}"), n, args.len()));
+        }
+
+        let comps = args
+            .iter()
+            .map(|a| {
+                a.convert_inner_to(tplt_ty)
+                    .ok_or_else(|| E::ParamType(tplt_ty.clone(), a.ty()))
+            })
+            .collect::<Result<Vec<_>, _>>()?;
+
+        Ok(VecInstance::new(comps).into())
+    }
+}
+
+/// `vecN()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#vec2-builtin>
+pub fn vec(n: usize, args: &[Instance]) -> Result<Instance, E> {
+    // overload 1: vec init from single scalar value
+    if let [Instance::Literal(l)] = args {
+        let val = Instance::Literal(*l);
+        let comps = (0..n).map(|_| val.clone()).collect_vec();
+        Ok(VecInstance::new(comps).into())
+    }
+    // overload 2: vec conversion constructor
+    else if let [Instance::Vec(v)] = args {
+        if v.n() != n {
+            let ty = v.ty();
+            let ty2 = Type::Vec(n as u8, ty.inner_ty().into());
+            return Err(E::Conversion(ty, ty2));
+        }
+        // note: `vecN(e: vecN<S>) -> vecN<S>` is no-op
+        Ok(v.clone().into())
+    }
+    // overload 3: vec init from component values
+    else if !args.is_empty() {
+        // flatten vecN args
+        let args = args
+            .iter()
+            .flat_map(|a| -> Box<dyn Iterator<Item = &Instance>> {
+                match a {
+                    Instance::Vec(v) => Box::new(v.iter()),
+                    _ => Box::new(std::iter::once(a)),
+                }
+            })
+            .cloned()
+            .collect_vec();
+        if args.len() != n {
+            return Err(E::ParamCount(format!("vec{n}"), n, args.len()));
+        }
+
+        let comps = convert_all(&args).ok_or(E::Builtin("vector components are incompatible"))?;
+
+        if !comps.first().unwrap(/* SAFETY: len() checked above */).ty().is_scalar() {
+            return Err(E::Builtin("vec constructor expects scalar arguments"));
+        }
+        Ok(VecInstance::new(comps).into())
+    }
+    // overload 3: zero-vec
+    else {
+        VecInstance::zero_value(n as u8, &Type::AbstractInt).map(Into::into)
+    }
+}
+
+/// User-defined struct constructor.
+pub fn struct_ctor(struct_ty: &StructType, args: &[Instance]) -> Result<StructInstance, E> {
+    if args.is_empty() {
+        return StructInstance::zero_value(struct_ty);
     }
 
     if args.len() != struct_ty.members.len() {
@@ -64,7 +555,7 @@ pub fn struct_ctor(struct_ty: &StructType, args: &[Instance]) -> Result<Instance
         })
         .collect::<Result<Vec<_>, E>>()?;
 
-    Ok(Instance::Struct(StructInstance::new(struct_ty.clone(), members)).into())
+    Ok(StructInstance::new(struct_ty.clone(), members))
 }
 
 /// Check a struct constructor call signature.
@@ -97,6 +588,9 @@ pub fn typecheck_struct_ctor(struct_ty: &StructType, args: &[Type]) -> Result<()
 // CONSTRUCTOR TYPES
 // -----------------
 
+/// Return type of `array<T,N>()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#array-builtin>
 fn array_ctor_ty_t(tplt_ty: &Type, tplt_n: usize, args: &[Type]) -> Result<Type, E> {
     if let Some(arg) = args.iter().find(|arg| !arg.is_convertible_to(tplt_ty)) {
         Err(E::Conversion(arg.clone(), tplt_ty.clone()))
@@ -105,11 +599,17 @@ fn array_ctor_ty_t(tplt_ty: &Type, tplt_n: usize, args: &[Type]) -> Result<Type,
     }
 }
 
+/// Return type of `array()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#array-builtin>
 fn array_ctor_ty(args: &[Type]) -> Result<Type, E> {
     let ty = convert_all_ty(args).ok_or(E::Builtin("array elements are incompatible"))?;
     Ok(Type::Array(Box::new(ty.clone()), Some(args.len())))
 }
 
+/// Return type of `matCxR<T>()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#mat2x2-builtin>
 fn mat_ctor_ty_t(c: u8, r: u8, tplt_ty: &Type, args: &[Type]) -> Result<Type, E> {
     // overload 1: mat conversion constructor
     if let [ty @ Type::Mat(c2, r2, _)] = args {
@@ -151,6 +651,9 @@ fn mat_ctor_ty_t(c: u8, r: u8, tplt_ty: &Type, args: &[Type]) -> Result<Type, E>
     Ok(Type::Mat(c, r, Box::new(tplt_ty.clone())))
 }
 
+/// Return type of `matCxR()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#mat2x2-builtin>
 fn mat_ctor_ty(c: u8, r: u8, args: &[Type]) -> Result<Type, E> {
     // overload 1: mat conversion constructor
     if let [ty @ Type::Mat(c2, r2, ty2)] = args {
@@ -191,6 +694,9 @@ fn mat_ctor_ty(c: u8, r: u8, args: &[Type]) -> Result<Type, E> {
     }
 }
 
+/// Return type of `vecN<T>()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#vec2-builtin>
 fn vec_ctor_ty_t(n: u8, tplt_ty: &Type, args: &[Type]) -> Result<Type, E> {
     if let [arg] = args {
         // overload 1: vec init from single scalar value
@@ -227,6 +733,9 @@ fn vec_ctor_ty_t(n: u8, tplt_ty: &Type, args: &[Type]) -> Result<Type, E> {
     Ok(Type::Vec(n, Box::new(tplt_ty.clone())))
 }
 
+/// Return type of `vecN()` constructor.
+///
+/// Reference: <https://www.w3.org/TR/WGSL/#vec2-builtin>
 fn vec_ctor_ty(n: u8, args: &[Type]) -> Result<Type, E> {
     if let [arg] = args {
         // overload 1: vec init from single scalar value
@@ -274,8 +783,8 @@ fn vec_ctor_ty(n: u8, args: &[Type]) -> Result<Type, E> {
 ///
 /// The arguments must be [loaded][Type::loaded].
 ///
-/// Includes built-in constructors and zero-value constructors, *except* the struct
-/// zero-value constructor, since it requires knowledge of the struct type.
+/// Includes built-in constructors and zero-value constructors, *but not* the struct
+/// constructors, since they require knowledge of the struct type.
 /// You can type-check a struct constructor call with [`typecheck_struct_ctor`].
 pub fn type_ctor(name: &str, tplt: Option<&[TpltParam]>, args: &[Type]) -> Result<Type, E> {
     match (name, tplt, args) {
